JavaBean update: Old 1.sql type tinyint -> java type Integer 2.sql type datetime -> java type sql.Date 3.sql type decimal -> java type Double new 1.sql type tinyint -> java type byte 2.sql type datetime -> java type sql.Timestamp 3.sql type decimal -> java type math.BigDecimal

+++ fallElove_springboot/src/main/java/JavaBean/rental/RentSet.java
@@ -3,29 +3,29 @@ package JavaBean.rental;
 public class RentSet implements java.io.Serializable{
     private Integer rOrdNo;
     private String rSetName;
-    private Integer rSetDays;
+    private Byte rSetDays;
 
-    public Integer getrOrdNo(){
+    public Integer getrOrdNo() {
         return rOrdNo;
     }
 
-    public void setrOrdNo(Integer rOrdNo){
+    public void setrOrdNo(Integer rOrdNo) {
         this.rOrdNo = rOrdNo;
     }
 
-    public String getrSetName(){
+    public String getrSetName() {
         return rSetName;
     }
 
-    public void setrSetName(String rSetName){
+    public void setrSetName(String rSetName) {
         this.rSetName = rSetName;
     }
 
-    public Integer getrSetDays(){
+    public Byte getrSetDays() {
         return rSetDays;
     }
 
-    public void setrSetDays(Integer rSetDays){
+    public void setrSetDays(Byte rSetDays) {
         this.rSetDays = rSetDays;
     }
 }
