+++ fallElove_springboot/src/main/java/JavaBean/rental/Rental.java
@@ -1,78 +1,78 @@
 package JavaBean.rental;
 
+import java.math.BigDecimal;
+
 public class Rental implements java.io.Serializable{
     private Integer rNo;
     private Integer rCatNo;
     private String rName;
-    private Double rPrice;
+    private BigDecimal rPrice;
     private Integer rSize;
     private String rColor;
     private String rInfo;
-    private Integer rStat;
+    private Byte rStat;
 
-
-    public Integer getrNo(){
+    public Integer getrNo() {
         return rNo;
     }
 
-    public void setrNo(Integer rNo){
+    public void setrNo(Integer rNo) {
         this.rNo = rNo;
     }
 
-
-    public Integer getrCatNo(){
+    public Integer getrCatNo() {
         return rCatNo;
     }
 
-    public void setrCatNo(Integer rCatNo){
+    public void setrCatNo(Integer rCatNo) {
         this.rCatNo = rCatNo;
     }
 
-    public String getrName(){
+    public String getrName() {
         return rName;
     }
 
-    public void setrName(String rName){
+    public void setrName(String rName) {
         this.rName = rName;
     }
 
-    public Double getrPrice(){
+    public BigDecimal getrPrice() {
         return rPrice;
     }
 
-    public void setrPrice(Double rPrice){
+    public void setrPrice(BigDecimal rPrice) {
         this.rPrice = rPrice;
     }
 
-    public Integer getrSize(){
+    public Integer getrSize() {
         return rSize;
     }
 
-    public void setrSize(Integer rSize){
+    public void setrSize(Integer rSize) {
         this.rSize = rSize;
     }
 
-    public String getrColor(){
+    public String getrColor() {
         return rColor;
     }
 
-    public void setrColor(String rColor){
+    public void setrColor(String rColor) {
         this.rColor = rColor;
     }
 
-    public String getrInfo(){
+    public String getrInfo() {
         return rInfo;
     }
 
-    public void setrInfo(String rInfo){
+    public void setrInfo(String rInfo) {
         this.rInfo = rInfo;
     }
 
-    public Integer getrStat(){
+    public Byte getrStat() {
         return rStat;
     }
 
-    public void setrStat(Integer rStat){
+    public void setrStat(Byte rStat) {
         this.rStat = rStat;
     }
 }
